# apiv6/localhost_apiserver/api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    name: Kentik API Engineering
    url: https://github.com/kentik/api
  title: Cloud Export Admin API
  version: "202101"
externalDocs:
  description: More about Kentik APIs
  url: https://docs.kentik.com/api
servers:
- url: /
security:
- email: []
  token: []
tags:
- name: CloudExportAdminService
paths:
  /cloud_export/v202101beta1/exports:
    get:
      operationId: CloudExportAdminService_ListCloudExport
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v202101beta1ListCloudExportResponse'
          description: A successful response.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/googlerpcStatus'
          description: An unexpected error response.
      tags:
      - CloudExportAdminService
    post:
      operationId: CloudExportAdminService_CreateCloudExport
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/v202101beta1CreateCloudExportRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v202101beta1CreateCloudExportResponse'
          description: A successful response.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/googlerpcStatus'
          description: An unexpected error response.
      tags:
      - CloudExportAdminService
  /cloud_export/v202101beta1/exports/{export.id}:
    delete:
      operationId: CloudExportAdminService_DeleteCloudExport
      parameters:
      - explode: false
        in: path
        name: export.id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v202101beta1DeleteCloudExportResponse'
          description: A successful response.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/googlerpcStatus'
          description: An unexpected error response.
      tags:
      - CloudExportAdminService
    get:
      operationId: CloudExportAdminService_GetCloudExport
      parameters:
      - explode: false
        in: path
        name: export.id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v202101beta1GetCloudExportResponse'
          description: A successful response.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/googlerpcStatus'
          description: An unexpected error response.
      tags:
      - CloudExportAdminService
    patch:
      operationId: CloudExportAdminService_PatchCloudExport
      parameters:
      - description: The internal cloud export identifier. This is Read-only and assigned
          by Kentik.
        explode: false
        in: path
        name: export.id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/v202101beta1PatchCloudExportRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v202101beta1PatchCloudExportResponse'
          description: A successful response.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/googlerpcStatus'
          description: An unexpected error response.
      tags:
      - CloudExportAdminService
    put:
      operationId: CloudExportAdminService_UpdateCloudExport
      parameters:
      - description: The internal cloud export identifier. This is Read-only and assigned
          by Kentik.
        explode: false
        in: path
        name: export.id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/v202101beta1UpdateCloudExportRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v202101beta1UpdateCloudExportResponse'
          description: A successful response.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/googlerpcStatus'
          description: An unexpected error response.
      tags:
      - CloudExportAdminService
components:
  schemas:
    cloud_exportv202101beta1Status:
      example:
        flowFound: true
        apiAccess: true
        storageAccountAccess: true
        errorMessage: errorMessage
        status: status
      properties:
        status:
          title: FIXME
          type: string
        errorMessage:
          title: If not empty, the current error
          type: string
        flowFound:
          title: If true, we found flow logs
          type: boolean
        apiAccess:
          type: boolean
        storageAccountAccess:
          type: boolean
      title: Export task status
      type: object
    googlerpcStatus:
      properties:
        code:
          format: int32
          type: integer
        message:
          type: string
        details:
          items:
            $ref: '#/components/schemas/protobufAny'
          type: array
      type: object
    protobufAny:
      properties:
        typeUrl:
          type: string
        value:
          format: byte
          type: string
      type: object
    v202101beta1AwsProperties:
      example:
        bucket: bucket
        multipleBuckets: true
        iamRoleArn: iamRoleArn
        deleteAfterRead: true
        region: region
      properties:
        bucket:
          title: Source S3 bucket to fetch vpc flow logs from
          type: string
        iamRoleArn:
          title: ARN for the IAM role to assume when fetching data or making AWS calls
            for this export
          type: string
        region:
          title: AWS region where this bucket resides (FIXME is that right?)
          type: string
        deleteAfterRead:
          title: If true, attempt to delete vpc flow log chunks from S3 after they've
            been read
          type: boolean
        multipleBuckets:
          title: FIXME
          type: boolean
      title: Properties specific to Amazon Web Services "vpc flow logs" exports
      type: object
    v202101beta1AzureProperties:
      example:
        resourceGroup: resourceGroup
        securityPrincipalEnabled: true
        location: location
        storageAccount: storageAccount
        subscriptionId: subscriptionId
      properties:
        location:
          type: string
        resourceGroup:
          type: string
        storageAccount:
          type: string
        subscriptionId:
          type: string
        securityPrincipalEnabled:
          type: boolean
      title: Properties specific to Azure exports
      type: object
    v202101beta1BgpProperties:
      description: Optional BGP related settings.
      example:
        deviceBgpType: deviceBgpType
        applyBgp: true
        useBgpDeviceId: useBgpDeviceId
      properties:
        applyBgp:
          description: If true, apply BGP data discovered via another device to the
            flow from this export.
          type: boolean
        useBgpDeviceId:
          title: Which other device to get BGP data from
          type: string
        deviceBgpType:
          title: 'FIXME: unclear. "device", "other_device" or "none"?'
          type: string
      type: object
    v202101beta1CloudExport:
      example:
        flowDest: flowDest
        currentStatus:
          flowFound: true
          apiAccess: true
          storageAccountAccess: true
          errorMessage: errorMessage
          status: status
        gce:
          project: project
          subscription: subscription
        description: description
        bgp:
          deviceBgpType: deviceBgpType
          applyBgp: true
          useBgpDeviceId: useBgpDeviceId
        enabled: true
        apiRoot: apiRoot
        ibm:
          bucket: bucket
        cloudProvider: cloudProvider
        name: name
        planId: planId
        id: id
        aws:
          bucket: bucket
          multipleBuckets: true
          iamRoleArn: iamRoleArn
          deleteAfterRead: true
          region: region
        azure:
          resourceGroup: resourceGroup
          securityPrincipalEnabled: true
          location: location
          storageAccount: storageAccount
          subscriptionId: subscriptionId
      properties:
        id:
          description: The internal cloud export identifier. This is Read-only and
            assigned by Kentik.
          type: string
        type:
          $ref: '#/components/schemas/v202101beta1CloudExportType'
        enabled:
          description: Whether this task is enabled and intended to run, or disabled.
          type: boolean
        name:
          description: A short name for this export.
          type: string
        description:
          description: An optional, longer description.
          type: string
        apiRoot:
          title: Hostname of the Kentik deployment where the API calls related to
            this export should go, eg. api.kentik.com
          type: string
        flowDest:
          title: Hostname of the Kentik deployment where the data generated by this
            export should go, eg. flow.kentik.com
          type: string
        planId:
          description: The identifier of the Kentik plan associated with this task.
          type: string
        cloudProvider:
          title: The cloud provider targeted by this export (eg. aws, azure, gce,
            ibm...)
          type: string
        aws:
          $ref: '#/components/schemas/v202101beta1AwsProperties'
        azure:
          $ref: '#/components/schemas/v202101beta1AzureProperties'
        gce:
          $ref: '#/components/schemas/v202101beta1GceProperties'
        ibm:
          $ref: '#/components/schemas/v202101beta1IbmProperties'
        bgp:
          $ref: '#/components/schemas/v202101beta1BgpProperties'
        currentStatus:
          $ref: '#/components/schemas/cloud_exportv202101beta1Status'
      title: The CloudExport message defines a cloud export task
      type: object
    v202101beta1CloudExportType:
      default: CLOUD_EXPORT_TYPE_UNSPECIFIED
      enum:
      - CLOUD_EXPORT_TYPE_UNSPECIFIED
      - CLOUD_EXPORT_TYPE_KENTIK_MANAGED
      - CLOUD_EXPORT_TYPE_CUSTOMER_MANAGED
      title: |-
        - CLOUD_EXPORT_TYPE_UNSPECIFIED: Invalid or incomplete exports.
         - CLOUD_EXPORT_TYPE_KENTIK_MANAGED: Cloud exports that are managed by Kentik.
         - CLOUD_EXPORT_TYPE_CUSTOMER_MANAGED: Exports that are managed by Kentik customers (eg. by running an agent)
      type: string
    v202101beta1CreateCloudExportRequest:
      properties:
        export:
          $ref: '#/components/schemas/v202101beta1CloudExport'
      type: object
    v202101beta1CreateCloudExportResponse:
      example:
        export:
          flowDest: flowDest
          currentStatus:
            flowFound: true
            apiAccess: true
            storageAccountAccess: true
            errorMessage: errorMessage
            status: status
          gce:
            project: project
            subscription: subscription
          description: description
          bgp:
            deviceBgpType: deviceBgpType
            applyBgp: true
            useBgpDeviceId: useBgpDeviceId
          enabled: true
          apiRoot: apiRoot
          ibm:
            bucket: bucket
          cloudProvider: cloudProvider
          name: name
          planId: planId
          id: id
          aws:
            bucket: bucket
            multipleBuckets: true
            iamRoleArn: iamRoleArn
            deleteAfterRead: true
            region: region
          azure:
            resourceGroup: resourceGroup
            securityPrincipalEnabled: true
            location: location
            storageAccount: storageAccount
            subscriptionId: subscriptionId
      properties:
        export:
          $ref: '#/components/schemas/v202101beta1CloudExport'
      type: object
    v202101beta1DeleteCloudExportResponse:
      type: object
    v202101beta1GceProperties:
      example:
        project: project
        subscription: subscription
      properties:
        project:
          type: string
        subscription:
          type: string
      title: Properties specific to Google Cloud export
      type: object
    v202101beta1GetCloudExportResponse:
      example:
        export:
          flowDest: flowDest
          currentStatus:
            flowFound: true
            apiAccess: true
            storageAccountAccess: true
            errorMessage: errorMessage
            status: status
          gce:
            project: project
            subscription: subscription
          description: description
          bgp:
            deviceBgpType: deviceBgpType
            applyBgp: true
            useBgpDeviceId: useBgpDeviceId
          enabled: true
          apiRoot: apiRoot
          ibm:
            bucket: bucket
          cloudProvider: cloudProvider
          name: name
          planId: planId
          id: id
          aws:
            bucket: bucket
            multipleBuckets: true
            iamRoleArn: iamRoleArn
            deleteAfterRead: true
            region: region
          azure:
            resourceGroup: resourceGroup
            securityPrincipalEnabled: true
            location: location
            storageAccount: storageAccount
            subscriptionId: subscriptionId
      properties:
        export:
          $ref: '#/components/schemas/v202101beta1CloudExport'
      type: object
    v202101beta1IbmProperties:
      example:
        bucket: bucket
      properties:
        bucket:
          type: string
      title: Properties specific to IBM Cloud exports
      type: object
    v202101beta1ListCloudExportResponse:
      example:
        invalidExportsCount: 0
        exports:
        - flowDest: flowDest
          currentStatus:
            flowFound: true
            apiAccess: true
            storageAccountAccess: true
            errorMessage: errorMessage
            status: status
          gce:
            project: project
            subscription: subscription
          description: description
          bgp:
            deviceBgpType: deviceBgpType
            applyBgp: true
            useBgpDeviceId: useBgpDeviceId
          enabled: true
          apiRoot: apiRoot
          ibm:
            bucket: bucket
          cloudProvider: cloudProvider
          name: name
          planId: planId
          id: id
          aws:
            bucket: bucket
            multipleBuckets: true
            iamRoleArn: iamRoleArn
            deleteAfterRead: true
            region: region
          azure:
            resourceGroup: resourceGroup
            securityPrincipalEnabled: true
            location: location
            storageAccount: storageAccount
            subscriptionId: subscriptionId
        - flowDest: flowDest
          currentStatus:
            flowFound: true
            apiAccess: true
            storageAccountAccess: true
            errorMessage: errorMessage
            status: status
          gce:
            project: project
            subscription: subscription
          description: description
          bgp:
            deviceBgpType: deviceBgpType
            applyBgp: true
            useBgpDeviceId: useBgpDeviceId
          enabled: true
          apiRoot: apiRoot
          ibm:
            bucket: bucket
          cloudProvider: cloudProvider
          name: name
          planId: planId
          id: id
          aws:
            bucket: bucket
            multipleBuckets: true
            iamRoleArn: iamRoleArn
            deleteAfterRead: true
            region: region
          azure:
            resourceGroup: resourceGroup
            securityPrincipalEnabled: true
            location: location
            storageAccount: storageAccount
            subscriptionId: subscriptionId
      properties:
        exports:
          items:
            $ref: '#/components/schemas/v202101beta1CloudExport'
          type: array
        invalidExportsCount:
          format: int64
          type: integer
      type: object
    v202101beta1PatchCloudExportRequest:
      properties:
        export:
          $ref: '#/components/schemas/v202101beta1CloudExport'
        updateMask:
          type: string
      type: object
    v202101beta1PatchCloudExportResponse:
      example:
        export:
          flowDest: flowDest
          currentStatus:
            flowFound: true
            apiAccess: true
            storageAccountAccess: true
            errorMessage: errorMessage
            status: status
          gce:
            project: project
            subscription: subscription
          description: description
          bgp:
            deviceBgpType: deviceBgpType
            applyBgp: true
            useBgpDeviceId: useBgpDeviceId
          enabled: true
          apiRoot: apiRoot
          ibm:
            bucket: bucket
          cloudProvider: cloudProvider
          name: name
          planId: planId
          id: id
          aws:
            bucket: bucket
            multipleBuckets: true
            iamRoleArn: iamRoleArn
            deleteAfterRead: true
            region: region
          azure:
            resourceGroup: resourceGroup
            securityPrincipalEnabled: true
            location: location
            storageAccount: storageAccount
            subscriptionId: subscriptionId
      properties:
        export:
          $ref: '#/components/schemas/v202101beta1CloudExport'
      type: object
    v202101beta1UpdateCloudExportRequest:
      properties:
        export:
          $ref: '#/components/schemas/v202101beta1CloudExport'
      type: object
    v202101beta1UpdateCloudExportResponse:
      example:
        export:
          flowDest: flowDest
          currentStatus:
            flowFound: true
            apiAccess: true
            storageAccountAccess: true
            errorMessage: errorMessage
            status: status
          gce:
            project: project
            subscription: subscription
          description: description
          bgp:
            deviceBgpType: deviceBgpType
            applyBgp: true
            useBgpDeviceId: useBgpDeviceId
          enabled: true
          apiRoot: apiRoot
          ibm:
            bucket: bucket
          cloudProvider: cloudProvider
          name: name
          planId: planId
          id: id
          aws:
            bucket: bucket
            multipleBuckets: true
            iamRoleArn: iamRoleArn
            deleteAfterRead: true
            region: region
          azure:
            resourceGroup: resourceGroup
            securityPrincipalEnabled: true
            location: location
            storageAccount: storageAccount
            subscriptionId: subscriptionId
      properties:
        export:
          $ref: '#/components/schemas/v202101beta1CloudExport'
      type: object
  securitySchemes:
    email:
      in: header
      name: X-CH-Auth-Email
      type: apiKey
    token:
      in: header
      name: X-CH-Auth-API-Token
      type: apiKey